enableIO undobugfix prefab quaternionbugfix physxbugfix

+++ TenshiEngine/Source/Game/EngineObject/ProfileViewer.cpp
@@ -50,8 +50,8 @@ public:
 
 		double time = Profiling::GetTime(mName);
 
-		float Width = 1200.0f;//(float)WindowState::mWidth;
-		float sx = (float)(time/(1.0 / 60.0)) * (Width) * 0.5;
+		double Width = 1200.0;//(float)WindowState::mWidth;
+		float sx = (float)((time/(1.0 / 60.0)) * (Width) * 0.5);
 
 		mTransform->Position(XMVectorSet(x + sx/2, pos.y, 0, 1));
 		mTransform->Scale(XMVectorSet(sx,scale.y, 1, 1));
